network/requirements.txt
inkbox
langchain
langchain-openai
flask
flask-cors
python-dotenv
openai
